+++ lib/strip.ps1
@@ -1,4 +1,5 @@
 class Strip : IRemote {
+    [System.Collections.ArrayList]$gainlayer
     [Object]$levels
 
     Strip ([int]$index, [Object]$remote) : base ($index, $remote) {
@@ -10,6 +11,11 @@ class Strip : IRemote {
         AddGainlayerMembers
 
         $this.levels = [StripLevels]::new($index, $remote)
+        
+        $this.gainlayer = @()
+        for ($i = 0; $i -lt $remote.kind.gainlayer; $i++) {
+            $this.gainlayer.Add([FloatArrayMember]::new($i, 'gainlayer', $this, 1))
+        }
     }
 
     [string] identifier () {